+++ tests/unit/components/shared/CodeEditor.spec.js
@@ -2,12 +2,13 @@ import Vuex from 'vuex'
 import CodeEditor from '@/components/shared/CodeEditor'
 import { createLocalVue, mount } from '@vue/test-utils'
 import { theme } from '@/store/modules/theme'
+import '../../mocks/codeEditorMocks'
 
 describe('components/shared/CodeEditor.vue', () => {
   let localVue
   let store
 
-  before(() => {
+  beforeEach(() => {
     localVue = createLocalVue()
     localVue.use(Vuex)
     store = new Vuex.Store({
@@ -16,27 +17,45 @@ describe('components/shared/CodeEditor.vue', () => {
   })
 
   it('renders when empty', () => {
-    const wrapper = mount(CodeEditor, {localVue, store})
-    expect(wrapper.html()).not.to.be.empty
+    const wrapper = mount(CodeEditor, {
+      localVue: localVue,
+      store: store,
+      propsData: {
+        useWorker: false
+      }
+    })
+    expect(wrapper.html()).not.toBeNull()
   })
 
   describe('theme handling', () => {
     it('uses the correct default theme', () => {
-      const wrapper = mount(CodeEditor, {localVue, store})
+      const wrapper = mount(CodeEditor, {
+        localVue: localVue,
+        store: store,
+        propsData: {
+          useWorker: false
+        }
+      })
       const vm = wrapper.vm
-      expect(vm.editor.getTheme()).to.eql('ace/theme/monokai')
+      expect(vm.editor.getTheme()).toEqual('ace/theme/monokai')
     })
 
     it('changes the editor theme correctly', () => {
-      const wrapper = mount(CodeEditor, {localVue, store})
+      const wrapper = mount(CodeEditor, {
+        localVue: localVue,
+        store: store,
+        propsData: {
+          useWorker: false
+        }
+      })
       const vm = wrapper.vm
-      expect(vm.editor.getTheme()).to.eql('ace/theme/monokai')
+      expect(vm.editor.getTheme()).toEqual('ace/theme/monokai')
 
       vm.$store.commit('theme/setDark', false)
-      expect(vm.editor.getTheme()).to.eql('ace/theme/textmate')
+      expect(vm.editor.getTheme()).toEqual('ace/theme/textmate')
 
       vm.$store.commit('theme/setDark', true)
-      expect(vm.editor.getTheme()).to.eql('ace/theme/monokai')
+      expect(vm.editor.getTheme()).toEqual('ace/theme/monokai')
     })
   })
 
@@ -46,36 +65,26 @@ describe('components/shared/CodeEditor.vue', () => {
         localVue: localVue,
         store: store,
         propsData: {
-          readOnly: true
+          readOnly: true,
+          useWorker: false
         }
       })
 
       const vm = wrapper.vm
-      expect(vm.editor.getReadOnly()).to.be.true
+      expect(vm.editor.getReadOnly()).toBeTruthy()
     })
-  })
-
-  describe('code handling', () => {
-    // it('changes the code when typing', () => {
-    //   let myCode = JSON.stringify({some: 'code'})
-    //   const wrapper = mount(CodeEditor, {
-    //     localVue: localVue,
-    //     store: store,
-    //     propsData: {
-    //       'code.sync': myCode
-    //     },
-    //     attachToDocument: true
-    //   })
-    //
-    //   // FIXME
-    //   expect(wrapper.find('.ace_text-layer').html()).to.include(myCode)
-    // })
   })
 
   describe('destroying', () => {
     it('destroys itself and the editor correctly', () => {
-      const wrapper = mount(CodeEditor, {localVue, store})
-      expect(wrapper.destroy())
+      const wrapper = mount(CodeEditor, {
+        localVue: localVue,
+        store: store,
+        propsData: {
+          useWorker: false
+        }
+      })
+      expect(wrapper.destroy()).toBeFalsy()
     })
   })
 })