+++ tests/unit/helpers/capitalize.spec.js
@@ -2,24 +2,24 @@ import { capitalize } from '@/helpers'
 
 describe('helpers/capitalize', () => {
   it('throws when called without a string', () => {
-    expect(() => capitalize()).to.throw(TypeError)
-    expect(() => capitalize(null)).to.throw(TypeError)
-    expect(() => capitalize(false)).to.throw(TypeError)
-    expect(() => capitalize(true)).to.throw(TypeError)
-    expect(() => capitalize(1)).to.throw(TypeError)
-    expect(() => capitalize([1])).to.throw(TypeError)
-    expect(() => capitalize({a: 1})).to.throw(TypeError)
+    expect(() => capitalize()).toThrow(TypeError)
+    expect(() => capitalize(null)).toThrow(TypeError)
+    expect(() => capitalize(false)).toThrow(TypeError)
+    expect(() => capitalize(true)).toThrow(TypeError)
+    expect(() => capitalize(1)).toThrow(TypeError)
+    expect(() => capitalize([1])).toThrow(TypeError)
+    expect(() => capitalize({a: 1})).toThrow(TypeError)
   })
 
   it('returns empty string when called with empty string', () => {
-    expect(capitalize('')).to.eql('')
+    expect(capitalize('')).toEqual('')
   })
 
   it('correctly capitalizes a word', () => {
-    expect(capitalize('string')).to.eql('String')
+    expect(capitalize('string')).toEqual('String')
   })
 
   it('correctly capitalizes a character', () => {
-    expect(capitalize('s')).to.eql('S')
+    expect(capitalize('s')).toEqual('S')
   })
 })