+++ tests/unit/helpers/filters.spec.js
@@ -21,42 +21,42 @@ describe('helpers/filters.js', () => {
   describe('filters.js default searches', () => {
     it('should return all items on empty search', () => {
       let results = fuzzyTableFilter(items, '', headers)
-      expect(results.length).to.equal(items.length)
+      expect(results.length).toEqual(items.length)
     })
 
     it('should return all items on whitespace search', () => {
       let results = fuzzyTableFilter(items, '   ', headers)
-      expect(results.length).to.equal(items.length)
+      expect(results.length).toEqual(items.length)
     })
 
     it('should return no items on unmatched search', () => {
       let results = fuzzyTableFilter(items, 'xxxxxxxxxxxxxx', headers)
-      expect(results).to.be.empty
+      expect(results).toEqual([])
     })
 
     it('should return all persons on person search', () => {
       const persons = items.filter(item => item.type === 'person')
-      const personNames = persons.map(person => person.name)
+      const personNames = persons.map(person => person.name).sort()
 
       const results = fuzzyTableFilter(items, 'person', headers)
-      const resultNames = results.map(result => result.name)
+      const resultNames = results.map(result => result.name).sort()
 
-      expect(results.length).to.equal(persons.length)
-      expect(resultNames).to.have.members(personNames)
+      expect(results.length).toEqual(persons.length)
+      expect(resultNames).toEqual(expect.arrayContaining(personNames))
     })
 
     it('should fuzzy match on "da" search', () => {
       const results = fuzzyTableFilter(items, 'da', headers)
       const resultNames = results.map(result => result.name)
       const expected = ['dagobah', 'darth vader', 'yoda']
-      expect(resultNames).to.have.members(expected)
+      expect(resultNames).toEqual(expect.arrayContaining(expected))
     })
 
     it('should fuzzy match on "dar" search', () => {
       const results = fuzzyTableFilter(items, 'dar', headers)
       const resultNames = results.map(result => result.name)
       const expected = ['darth vader']
-      expect(resultNames).to.have.members(expected)
+      expect(resultNames).toEqual(expect.arrayContaining(expected))
     })
 
     it('should filter all things with s on s search', () => {
@@ -65,19 +65,19 @@ describe('helpers/filters.js', () => {
       const results = fuzzyTableFilter(items, 's', headers)
       const resultNames = results.map(result => result.name)
 
-      expect(resultNames).to.have.members(expected)
+      expect(resultNames).toEqual(expect.arrayContaining(expected))
     })
   })
 
   describe('filters.js column searches', () => {
     it('should return all items on empty column search', () => {
       const results = fuzzyTableFilter(items, 'type:', headers)
-      expect(results.length).to.equal(items.length)
+      expect(results.length).toEqual(items.length)
     })
 
     it('should return all items on whitespace column search', () => {
       const results = fuzzyTableFilter(items, 'type:   ', headers)
-      expect(results.length).to.equal(items.length)
+      expect(results.length).toEqual(items.length)
     })
 
     it('should return all persons on type:person search', () => {
@@ -87,8 +87,8 @@ describe('helpers/filters.js', () => {
       const results = fuzzyTableFilter(items, 'type:person', headers)
       const resultNames = results.map(result => result.name)
 
-      expect(results.length).to.equal(persons.length)
-      expect(resultNames).to.have.members(personNames)
+      expect(results.length).toEqual(persons.length)
+      expect(resultNames).toEqual(expect.arrayContaining(personNames))
     })
 
     it('should filter all ships and persons on type:s search', () => {
@@ -98,7 +98,7 @@ describe('helpers/filters.js', () => {
       const results = fuzzyTableFilter(items, 'type:s', headers)
       const resultNames = results.map(result => result.name)
 
-      expect(resultNames).to.have.members(itemNames)
+      expect(resultNames).toEqual(expect.arrayContaining(itemNames))
     })
   })
 })