+++ tests/unit/helpers/utilities/flattenObject.spec.js
@@ -3,41 +3,41 @@ import { flattenObject } from '@/helpers/utilities'
 describe('helpers/utilities/flattenObject', () => {
   describe('flattenObject with empty object', () => {
     it('returns empty object for empty object without deep and concatNames', () => {
-      expect(flattenObject({})).to.eql({})
+      expect(flattenObject({})).toEqual({})
     })
 
     it('returns empty array for empty object with deep and without concatNames', () => {
-      expect(flattenObject({}, true)).to.eql({})
+      expect(flattenObject({}, true)).toEqual({})
     })
 
     it('returns empty array for empty object with deep concatNames', () => {
-      expect(flattenObject({}, true, true)).to.eql({})
+      expect(flattenObject({}, true, true)).toEqual({})
     })
 
     it('returns empty array for empty object without deep and with concatNames', () => {
-      expect(flattenObject({}, false, true)).to.eql({})
+      expect(flattenObject({}, false, true)).toEqual({})
     })
   })
 
   describe('utilities/flattenObject without nesting', () => {
     it('returns the already flattened object when called without options', () => {
       const obj = {a: 1, b: 2}
-      expect(flattenObject(obj)).to.eql(obj)
+      expect(flattenObject(obj)).toEqual(obj)
     })
 
     it('returns the already flattened object when called with deep option', () => {
       const obj = {a: 1, b: 2}
-      expect(flattenObject(obj, true)).to.eql(obj)
+      expect(flattenObject(obj, true)).toEqual(obj)
     })
 
     it('returns the already flattened object when called with concat option', () => {
       const obj = {a: 1, b: 2}
-      expect(flattenObject(obj, false, true)).to.eql(obj)
+      expect(flattenObject(obj, false, true)).toEqual(obj)
     })
 
     it('returns the already flattened object when called with deep and concat option', () => {
       const obj = {a: 1, b: 2}
-      expect(flattenObject(obj, true, true)).to.eql(obj)
+      expect(flattenObject(obj, true, true)).toEqual(obj)
     })
   })
 
@@ -45,25 +45,25 @@ describe('helpers/utilities/flattenObject', () => {
     it('returns the flattened object when called without options', () => {
       const obj = {a: 1, b: 2, c: {d: 1}}
       const flattened = {a: 1, b: 2, d: 1}
-      expect(flattenObject(obj)).to.eql(flattened)
+      expect(flattenObject(obj)).toEqual(flattened)
     })
 
     it('returns the first level flattened object when called without options', () => {
       const obj = {a: 1, b: 2, c: {d: 1, e: {f: 3}}}
       const flattened = {a: 1, b: 2, d: 1, e: {f: 3}}
-      expect(flattenObject(obj)).to.eql(flattened)
+      expect(flattenObject(obj)).toEqual(flattened)
     })
 
     it('returns the deep flattened object when called with deep', () => {
       const obj = {a: 1, b: 2, c: {d: 1, e: {f: 3}}}
       const flattened = {a: 1, b: 2, d: 1, f: 3}
-      expect(flattenObject(obj, true)).to.eql(flattened)
+      expect(flattenObject(obj, true)).toEqual(flattened)
     })
 
     it('returns the concatenated deep flattened object when called with deep and concat', () => {
       const obj = {a: 1, b: 2, c: {d: 1, e: {f: 3}}}
       const flattened = {a: 1, b: 2, 'c.d': 1, 'c.e.f': 3}
-      expect(flattenObject(obj, true, true)).to.eql(flattened)
+      expect(flattenObject(obj, true, true)).toEqual(flattened)
     })
   })
 })